Create upvotes for comments and posts
1. Make an arrow component for both posts and comments
2. Make api calls to backend to create an upvote or downvote
3. On the backend if a user already exists in either upvote or downvote,
depending on if I want to upvote or downvote, then return 400. If a
400 status was returned delete said upvote or downvote.
4. If a 201 is returned create that upvote or downvote.
5. If an upvote or downvote already existed when pressed the button,
destroy and create said upvote or downvote
6. This works for both posts and comments the same way, but I have
separated them out in the backend. It maybe possible to mix them
together in order to not repeat myself

UnAuthPost and posts I added cleanup functions to the end of the
useEffect