Maintain vote status
The main premise is that I create a post or comment with an array of
upvote users/downvote users. If there is an upvote user and they match
the current signed in user make an upvote, if one in downvote then make
it downvote. When pushing the increment or decrement buttons I handle
storing the status in the backend. See backend commit for more info.
Back on the front end I make visual changes by passing posts down from
routes.js. I make changes to the post that I am working on by passing
a as a parameter what I am doing to the vote. Its similar to the backend
in this regard.