Fix bug with comments
Whenever adding or deleting a comment happened I was making every
post update the comment amount to be the same. Fixed by doing a loop
though the posts and fining the post id and updating the amount through
this function. This is very similar to how I solved the upvote problem
before. I also updated css to be more responsive for comments